Refacts AnswerView Page (#18)
* feat(AnswerPage):
 - refacts the QuestionDetail page transforming it at AnswerPage with better use of the componentizing
 - creates a StatusMessageProvider to handle general messages of the app
 - creates a AnswersMessageProvider to handle shared answers' state between Answer page's components
 - creates a StakingProvider to provide some staking states through the Answer page's components

* refact(QuestionPage):
 - changes the NewQuestion page's name to QuestionPage

* feat(useWallet, useStatusMessage):
 - reorganizes the useWallet context, provider and hook, in separated files to avoid conflits with React Fast-Refreshing.
 - creates the useStatusMessage hook.

* feat(Hero):
 - adds link to the app in the landing page

* fix(imports):
 - fix imports

* fix(ReactMarkdown):
 - it fix the import of the ReactMarkdown component
 - adjusts a build config in the vite.config

* fix(defaultTheme):
 - adds types for typescript default theme

* fix(theme): update hover, warning, and secondary properties in light and dark themes

* fix(DropdownItem): correct hover background property reference
