+++ frontend/react/src/@types/styles.d.ts
@@ -4,5 +4,9 @@ import { defaultTheme } from "../styles/themes";
 type ThemeType = typeof defaultTheme;
 
 declare module "styled-components" {
-  export interface DefaultTheme extends ThemeType {}
+  export interface DefaultTheme extends ThemeType {
+    hover: string
+    warning: string
+    secondary: string
+  }
 }
