+++ frontend/react/src/App.tsx
@@ -1,14 +1,15 @@
 import { ThemeProvider } from "styled-components"
 import { GlobalStyle } from "./styles/global"
-import { BrowserRouter } from "react-router-dom";
-import { Router } from "./router";
-import { Header } from "./components/Header";
-import { darkTheme, lightTheme } from "./styles/themes";
-import { WalletProvider } from "./providers/wallet-connect-context";
-import { WalletDetector } from "./components/wallet-detector";
-import { useState, useEffect } from "react";
+import { BrowserRouter } from "react-router-dom"
+import { Router } from "./router"
+import { Header } from "./components/Header"
+import { darkTheme, lightTheme } from "./styles/themes"
+import { WalletDetector } from "./components/wallet-detector"
+import { useState, useEffect } from "react"
+import { StatusMessageProvider } from "@hooks/useStatusMessage/statusMessage.provider"
+import { WalletProvider } from "@hooks/useWallet/wallet.provider"
 
-function App() {
+export function App() {
   const [theme, setTheme] = useState(darkTheme);  
   
   // Try to load preferred theme from localStorage
@@ -29,16 +30,16 @@ function App() {
 
   return (
     <ThemeProvider theme={theme}>
-    <WalletProvider>
-      <BrowserRouter>
-        <Header toggleTheme={toggleTheme}/>
-        <Router />
-        <WalletDetector />
-      </BrowserRouter>
-      <GlobalStyle />
-    </WalletProvider>
+    <StatusMessageProvider>
+      <WalletProvider>
+        <BrowserRouter>
+          <Header toggleTheme={toggleTheme}/>
+          <WalletDetector />
+          <Router />
+        </BrowserRouter>
+        <GlobalStyle />
+      </WalletProvider>
+    </StatusMessageProvider>
   </ThemeProvider>
   )
 }
-
-export default App
