+++ frontend/react/src/hooks/useStatusMessage/types.ts
@@ -1,0 +1,1 @@
+export type StatusMessage = { type: "success" | "error" | "info" | null; message: string }
