+++ frontend/react/src/hooks/useWallet/wallet.provider.tsx
@@ -1,21 +1,7 @@
-import React, { createContext, useContext, ReactNode, useState, useEffect } from "react";
+import { ReactNode, useEffect, useState } from "react";
+import { WalletContext } from "./wallet.context";
 import { useAccount, useConnect } from "@starknet-react/core";
 
-// Define the context type
-interface WalletContextType {
-  isConnected: boolean;
-  address: string | undefined;
-  isConnecting: boolean;
-  isWalletDetected: boolean; // Expose wallet detection state
-  isModalOpen: boolean;
-  openConnectModal: () => void;
-  closeConnectModal: () => void;
-}
-
-// Create the context
-const WalletContext = createContext<WalletContextType | undefined>(undefined);
-
-// Provider component
 export function WalletProvider({ children }: { children: ReactNode }) {
   const { isConnected = false, address, status } = useAccount();
   const { connectors } = useConnect();
@@ -55,13 +41,4 @@ export function WalletProvider({ children }: { children: ReactNode }) {
       {children}
     </WalletContext.Provider>
   );
-}
-
-// Custom hook to use the wallet context
-export function useWallet() {
-  const context = useContext(WalletContext);
-  if (context === undefined) {
-    throw new Error("useWallet must be used within a WalletProvider");
-  }
-  return context;
 }
