+++ frontend/react/src/pages/Forum/style.ts
@@ -1,4 +1,4 @@
-import { darken, rgba } from "polished";
+import { rgba } from "polished";
 import styled from "styled-components";
 
 export const ForumContainer = styled.div`
