+++ frontend/react/src/pages/QuestionPage/EditorForm/Toolbar/style.ts
