+++ frontend/react/src/providers/StarknetProvider.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { useEffect } from "react";
+import React from "react";
 import { mainnet } from "@starknet-react/chains";
 import {
   StarknetConfig,
