+++ frontend/react/src/styles/themes/dark.tsx
@@ -1,4 +1,4 @@
-export const darkTheme = {
+export const darkTheme = {  
   background: "#121212",
   text: "#e4e4e4",
   cardBackground: "#1e1e1e",
@@ -6,6 +6,9 @@ export const darkTheme = {
   connectBg: "#e4e4e4",
   connectText: "#121212",
   textSecondary: "#71717a",
+  hover: "#000",
+  warning: "#f39c12",
+  secondary: "#1e1e1e",
   forum: {
     topicCard: {
       background: "#1e1e1e",
