+++ frontend/react/tsconfig.json
@@ -27,6 +27,9 @@
     "paths": {
       "@logos/*": ["assets/logos/*"],
       "@components/*": ["components/*"],
+      "@providers/*": ["providers/*"],
+      "@utils/*": ["utils/*"],
+      "@hooks/*": ["hooks/*"],
     }
   },
   "include": ["src"]
