+++ frontend/react/vite.config.ts
@@ -5,10 +5,16 @@ import { fileURLToPath } from 'url';
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  build: {
+    target: "esnext",
+  },
   resolve: {
     alias: {
       "@logos": fileURLToPath(new URL("src/assets/logos", import.meta.url)),
       "@components": fileURLToPath(new URL("src/components", import.meta.url)),
+      "@providers": fileURLToPath(new URL("src/providers", import.meta.url)),
+      "@utils": fileURLToPath(new URL("src/utils", import.meta.url)),
+      "@hooks": fileURLToPath(new URL("src/hooks", import.meta.url)),
     },
   },
 });
